+++ modules/genrams/Manifest.py
@@ -1,18 +1,22 @@
-
 files = [
 	"genram_pkg.vhd",
 	"memory_loader_pkg.vhd"];
 
+# Be sure 'target' is defined.
 try:
         target
 except NameError:
         target = ""
 
-if (target == "altera"):
-	modules = {"local" : ["altera", "generic", "common"]}
-elif (target == "xilinx" and syn_device[0:4].upper()=="XC6V"):
-	modules = {"local" : ["xilinx", "xilinx/virtex6", "common"]}
-elif (target == "xilinx"):
-	modules = {"local" : ["xilinx", "generic", "common"]}
+# Target specific modules.
+if target == "altera":
+	local_modules = ["altera", "generic"]
+elif target == "xilinx":
+    if syn_device[0:4].upper()=="XC6V":
+        local_modules = ["xilinx", "xilinx/virtex6"]
+    else:
+        local_modules = ["xilinx", "generic"]
 else:
-	modules = {"local" : ["generic", "common"]}
+	local_modules = ["generic"]
+
+modules = {"local" : local_modules + ["common", "cheby"]}