+++ modules/genrams/cheby/Manifest.py
@@ -1,0 +1,3 @@
+files = [
+    "cheby_dpssram.vhd"
+]